requirements.txt
streamlit==1.4.0
snowflake-connector-python==2.6.0